# Improve performance for schemas with many fields