@Motiejus 2017-01-21
Dear maintainers,

We profiled graphql-go in our environment, and saw some performance issues associated to a growing schema.

What needs to happen to bring this to the finish line? Would it be easier for you to merge if the commits were merged to more small pull requests? Or is there something else we could do?