@fd 2018-01-08
This can be closed now?

See #267