@sfriedel 2018-01-24
closing this one since #267 addressed the major issue i saw (invariant/printf overhead)